use </blockquote> to find end of raidlog zone section
The raidlog order of zones doesn't appear to be consistent enough to depend on. (Mine showed up as Woods, Village, Miscellaneous, Castle, and the script was assuming it should be Village, Woods, Castle, Miscellaneous.)

Searching for `</blockquote>` starting from the beginning of the section finds the end of the section no matter where it shows up.